fix chat list

+++ api/src/dtos/chats.rs
@@ -9,6 +9,6 @@ pub struct ChatResponse {
     pub nickname: String,
     pub last_message: Option<String>,
     pub last_message_time: Option<DateTime<Utc>>,
-    pub unread: i32,
+    pub unread: i64,
     pub avatar_url: Option<String>,
 }
